@@ -27,14 +27,14 @@
     "dependencies": {
         "body-parser": "1.19.0",
         "bufferutil": "4.0.5",
-        "connect-session-knex": "2.1.0",
+        "connect-session-knex": "2.1.1",
         "csurf": "1.11.0",
-        "csv-parse": "4.16.3",
+        "csv-parse": "5.0.3",
         "express": "4.17.1",
         "express-session": "1.17.2",
         "express-zip": "3.0.0",
         "form-data": "4.0.0",
-        "knex": "0.95.12",
+        "knex": "0.95.14",
         "moment": "2.29.1",
         "multer": "1.4.3",
         "nodeify": "1.0.1",
@@ -42,39 +42,39 @@
         "passport": "0.5.0",
         "path": "0.12.7",
         "sqlite3": "5.0.2",
-        "swagger-ui-express": "4.1.6",
+        "swagger-ui-express": "4.2.0",
         "utf-8-validate": "5.0.7",
         "uuid": "8.3.2",
-        "ws": "8.2.3"
+        "ws": "8.3.0"
     },
     "devDependencies": {
-        "@babel/core": "7.15.8",
-        "@babel/eslint-parser": "7.15.8",
+        "@babel/core": "7.16.0",
+        "@babel/eslint-parser": "7.16.3",
         "@babel/eslint-plugin": "7.14.5",
         "@babel/plugin-syntax-dynamic-import": "7.8.3",
-        "@babel/preset-env": "7.15.8",
-        "@stryker-mutator/core": "5.4.1",
+        "@babel/preset-env": "7.16.4",
+        "@stryker-mutator/core": "5.5.1",
         "@stryker-mutator/html-reporter": "3.1.0",
         "@stryker-mutator/javascript-mutator": "4.0.0",
-        "@stryker-mutator/jest-runner": "5.4.1",
+        "@stryker-mutator/jest-runner": "5.5.1",
         "babel-jest": "26.6.3",
         "babel-loader": "8.2.3",
         "babel-plugin-require-context-hook": "1.0.0",
         "cross-env": "7.0.3",
-        "eslint": "7.32.0",
-        "eslint-plugin-promise": "5.1.1",
+        "eslint": "8.4.1",
+        "eslint-plugin-promise": "5.2.0",
         "jest": "26.6.3",
         "native-ext-loader": "2.3.0",
         "rimraf": "3.0.2",
         "start-server-webpack-plugin": "2.2.5",
-        "string-replace-loader": "3.0.3",
+        "string-replace-loader": "3.1.0",
         "supertest": "6.1.6",
-        "webpack": "4.46.0",
+        "webpack": "5.65.0",
         "webpack-cli": "4.9.1",
         "webpack-node-externals": "3.0.0"
     },
     "optionalDependencies": {
-        "ajv": "8.6.3",
+        "ajv": "8.8.2",
         "fsevents": "2.3.2"
     },
     "jest": {
