@@ -117,20 +117,20 @@
     },
     "dependencies": {
         "electron-log": "4.4.1",
-        "electron-updater": "4.3.9",
+        "electron-updater": "4.6.1",
         "express": "4.17.1",
         "sqlite3": "5.0.2"
     },
     "devDependencies": {
-        "@babel/core": "7.15.8",
+        "@babel/core": "7.16.0",
         "babel-loader": "8.2.3",
-        "copy-webpack-plugin": "6.4.1",
-        "electron": "14.2.0",
-        "electron-builder": "22.13.1",
+        "copy-webpack-plugin": "10.0.0",
+        "electron": "16.0.4",
+        "electron-builder": "22.14.5",
         "electron-reload": "1.5.0",
-        "html-webpack-plugin": "4.5.2",
-        "webpack": "4.46.0",
-        "webpack-command": "0.5.1",
+        "html-webpack-plugin": "5.5.0",
+        "webpack": "5.65.0",
+        "webpack-cli": "4.9.1",
         "webpack-merge": "5.8.0"
     }
 }
