@@ -24,32 +24,32 @@
         "csurf": "1.11.0",
         "express": "4.17.1",
         "express-session": "1.17.2",
-        "connect-mongo": "3.2.0",
-        "mongodb": "4.1.3",
+        "connect-mongo": "4.6.0",
+        "mongodb": "4.2.1",
         "passport": "0.5.0",
         "saslprep": "1.0.3"
     },
     "devDependencies": {
-        "@babel/core": "7.15.8",
-        "@babel/eslint-parser": "7.15.8",
+        "@babel/core": "7.16.0",
+        "@babel/eslint-parser": "7.16.3",
         "@babel/eslint-plugin": "7.14.5",
         "@babel/plugin-syntax-dynamic-import": "7.8.3",
-        "@babel/plugin-transform-modules-commonjs": "7.15.4",
-        "@babel/preset-env": "7.15.8",
+        "@babel/plugin-transform-modules-commonjs": "7.16.0",
+        "@babel/preset-env": "7.16.4",
         "babel-jest": "26.6.3",
         "babel-loader": "8.2.3",
         "babel-plugin-require-context-hook": "1.0.0",
         "cross-env": "7.0.3",
-        "eslint": "7.32.0",
-        "eslint-plugin-promise": "5.1.1",
+        "eslint": "8.4.1",
+        "eslint-plugin-promise": "5.2.0",
         "jest": "26.6.3",
         "rimraf": "3.0.2",
         "start-server-webpack-plugin": "2.2.5",
-        "webpack": "4.46.0",
+        "webpack": "5.65.0",
         "webpack-cli": "4.9.1"
     },
     "optionalDependencies": {
-        "ajv": "8.6.3",
+        "ajv": "8.8.2",
         "fsevents": "2.3.2"
     },
     "jest": {
