@@ -12,13 +12,13 @@
     },
     "private": true,
     "scripts": {
-        "start": "cross-env BROWSER=none SKIP_PREFLIGHT_CHECK=true react-scripts start",
-        "build": "cross-env SKIP_PREFLIGHT_CHECK=true react-scripts build",
-        "test": "cross-env SKIP_PREFLIGHT_CHECK=true CI=true react-scripts test --colors"
+        "start": "cross-env BROWSER=none SKIP_PREFLIGHT_CHECK=true react-app-rewired start",
+        "build": "cross-env SKIP_PREFLIGHT_CHECK=true react-app-rewired build",
+        "test": "cross-env SKIP_PREFLIGHT_CHECK=true CI=true react-app-rewired test --colors"
     },
     "dependencies": {
         "deepmerge": "4.2.2",
-        "detect-browser": "5.2.1",
+        "detect-browser": "5.3.0",
         "draft-js": "0.11.7",
         "file-saver": "2.0.5",
         "history": "4.10.1",
@@ -29,40 +29,43 @@
         "moment": "2.29.1",
         "react": "17.0.2",
         "react-calendar-timeline": "0.26.7",
-        "react-datepicker": "4.3.0",
+        "react-datepicker": "4.5.0",
         "react-dom": "17.0.2",
         "react-event-timeline": "1.6.3",
-        "react-helmet-async": "1.1.2",
-        "react-markdown": "5.0.3",
+        "react-helmet-async": "1.2.2",
+        "react-markdown": "7.1.1",
         "react-medium-image-zoom": "3.1.2",
         "react-redux": "7.2.6",
         "react-router-dom": "5.3.0",
         "react-virtualized": "9.22.3",
         "react-virtualized-tree": "3.4.1",
         "redux": "4.1.2",
-        "redux-thunk": "2.4.0"
+        "redux-thunk": "2.4.1"
     },
     "devDependencies": {
-        "@babel/core": "7.15.8",
-        "@babel/eslint-parser": "7.15.8",
+        "@babel/core": "7.16.0",
+        "@babel/eslint-parser": "7.16.3",
         "@babel/eslint-plugin": "7.14.5",
-        "@babel/plugin-proposal-decorators": "7.15.8",
-        "@babel/plugin-syntax-jsx": "7.14.5",
-        "@typescript-eslint/eslint-plugin": "4.33.0",
-        "@typescript-eslint/parser": "4.33.0",
+        "@babel/helper-builder-react-jsx": "7.16.0",
+        "@babel/helper-builder-react-jsx-experimental": "7.12.11",
+        "@babel/plugin-proposal-decorators": "7.16.4",
+        "@babel/plugin-syntax-jsx": "7.16.0",
+        "@typescript-eslint/eslint-plugin": "5.6.0",
+        "@typescript-eslint/parser": "5.6.0",
         "cross-env": "7.0.3",
-        "eslint": "7.32.0",
+        "eslint": "8.4.1",
         "eslint-config-react-app": "6.0.0",
-        "eslint-plugin-flowtype": "5.9.1",
-        "eslint-plugin-import": "2.25.2",
-        "eslint-plugin-jsx-a11y": "6.4.1",
-        "eslint-plugin-react": "7.26.1",
-        "eslint-plugin-react-hooks": "4.2.0",
+        "eslint-plugin-flowtype": "8.0.3",
+        "eslint-plugin-import": "2.25.3",
+        "eslint-plugin-jsx-a11y": "6.5.1",
+        "eslint-plugin-react": "7.27.1",
+        "eslint-plugin-react-hooks": "4.3.0",
         "http-proxy-middleware": "2.0.1",
         "markdown-image-loader": "2.0.1",
         "prop-types": "15.7.2",
-        "react-scripts-rewired": "3.1.1",
-        "webpack": "4.46.0"
+        "react-app-rewired": "2.1.8",
+        "react-scripts": "3.4.4",
+        "webpack": "5.65.0"
     },
     "eslintConfig": {
         "extends": "react-app"
