@@ -1,4 +1,6 @@
-import { tableMove } from '../utils/db-mover';
+import {
+    tableMove
+} from '../utils/db-mover';
 import ceFields from './defaults_v1/ceFields.json';
 import v7_ceFields from './defaults_v7/ceFields.json';
 import v8_ceFields from './defaults_v8/ceFields.json';
@@ -32,9 +34,18 @@ export default async (dbcon, version) => {
             await dbcon()(TABLE_NAME).insert(v7_ceFields);
             break;
         case 8:
-            await dbcon()(TABLE_NAME).where('idname', 'Severity').update({ permittedValues: 'Mild,Moderate,Severe,Unknown' });
-            await dbcon()(TABLE_NAME).where('idname', 'Recovery').update({ permittedValues: 'Complete,Partial,None,Unknown' });
+            await dbcon()(TABLE_NAME).where('idname', 'Severity').update({
+                permittedValues: 'Mild,Moderate,Severe,Unknown'
+            });
+            await dbcon()(TABLE_NAME).where('idname', 'Recovery').update({
+                permittedValues: 'Complete,Partial,None,Unknown'
+            });
             await dbcon()(TABLE_NAME).insert(v8_ceFields);
+            break;
+        case 13:
+            await dbcon()(TABLE_NAME).where('idname', 'Opportunistic Infection Classification').update({
+                permittedValues: 'PML,Herpes Zoster,Herpes Simplex,Varicella,Viral Hepatitis,Bacterial infection - listeria,Other infection,Mycosis,Abscess,COVID-19,Other'
+            });
             break;
         default:
             break;
