@@ -1,4 +1,7 @@
-import { tableMove, tableCopyBack } from '../utils/db-mover';
+import {
+    tableMove,
+    tableCopyBack
+} from '../utils/db-mover';
 
 export const TABLE_NAME = 'CONCOMITANT_MED';
 export const PRIORITY = 3;
@@ -17,6 +20,21 @@ export default async (dbcon, version) => {
                 table.integer('createdByUser').notNullable().references('id').inTable('USERS');
                 table.text('deleted').notNullable().defaultTo('-');
                 table.unique(['visit', 'concomitantMedId', 'deleted'], `UNIQUE_${Date.now()}_${TABLE_NAME}`);
+            });
+            await tableCopyBack(TABLE_NAME);
+            break;
+        case 13:
+            await tableMove(TABLE_NAME, version);
+            await dbcon().schema.createTable(TABLE_NAME, (table) => {
+                table.increments('id').primary();
+                table.integer('visit').notNullable().references('id').inTable('VISITS').onDelete('CASCADE');
+                table.integer('concomitantMedId').notNullable().references('id').inTable('AVAILABLE_CONCOMITANT_MED').onDelete('CASCADE');
+                table.text('indication').notNullable().defaultTo('Not provided');
+                table.text('createdTime').notNullable().defaultTo(dbcon().fn.now());
+                table.text('startDate').notNullable();
+                table.text('endDate').nullable();
+                table.integer('createdByUser').notNullable().references('id').inTable('USERS');
+                table.text('deleted').notNullable().defaultTo('-');
             });
             await tableCopyBack(TABLE_NAME);
             break;
